rename generate-config => generate-configs
