+++ lib/js/_create.js
@@ -8,7 +8,7 @@ const inquirer = require('inquirer');
 const Listr = require('listr');
 
 const exitWithError = require('./_utils/_error');
-const generateConfig = require('./_utils/_generate-config');
+const generateConfigs = require('./_utils/_generate-configs');
 const getLicense = require('./_utils/_get-license');
 const getPackages = require('./_utils/_get-packages');
 const getPackageName = require('./_utils/_get-package-name');
@@ -61,7 +61,7 @@ module.exports = (packageJsonPath, allToolkitsInfo) => {
 				.reduce((object, key) => ({...object, [key]: allToolkitsInfo[key]}), {});
 
 			// Add the config for selected toolkit
-			toolkitDetails = await generateConfig(packageJsonPath, toolkitDetails);
+			toolkitDetails = await generateConfigs(packageJsonPath, toolkitDetails);
 
 			// Add the package names for selected toolkit
 			// And save out the details of this toolkit
