+++ lib/js/_utils/_generate-configs.js
